select baseline calculation function and delete additional functions file
