+++ src/miniml/core/functions.py
@@ -1,6 +1,9 @@
+from collections.abc import Sequence
 from typing import NamedTuple
 
 import numpy as np
+import ruptures as rpt
+import scipy as sc
 
 
 class BaselineResult(NamedTuple):
@@ -70,7 +73,7 @@ def get_event_peak(
     return peak_position
 
 
-def get_event_baseline(
+def legacy_get_event_baseline(
     data: np.ndarray,
     duration: int,
     event_num: int,
@@ -156,6 +159,176 @@ def get_event_baseline(
 
     return BaselineResult(
         value=baseline, var=bsl_var, start=bsl_start, end=bsl_end, duration=bsl_duration
+    )
+
+
+def get_segment_stats(breakpoints: list, data: np.ndarray):
+    values, slopes, variances = [], [], []
+    for i, p2 in enumerate(breakpoints):
+        p1 = breakpoints[i - 1] if i else 0
+        p1 += 1
+        p2 -= 1
+        values.append(np.median(data[p1:p2]))
+        variances.append(np.std(data[p1:p2]))
+        if p2 - p1 > 1:
+            coef = (
+                np.polynomial.polynomial.Polynomial.fit(
+                    np.arange(p1, p2), data[p1:p2], 1
+                )
+                .convert()
+                .coef
+            )
+            if len(coef) > 1:
+                slopes.append(coef[1])
+            else:
+                slopes.append(0.0)
+        else:
+            slopes.append(0.0)
+
+    return np.array(values), np.array(variances), np.array(slopes)
+
+
+def get_steepest_rise_position(data: np.ndarray, filter_win: int = 20):
+    win = sc.signal.windows.hann(filter_win)
+    filtered_data = sc.signal.convolve(data, win, mode="same") / sum(win)
+
+    return np.argmax(np.gradient(filtered_data))
+
+
+def baseline_score(
+    positions: np.ndarray | list[int],
+    median_values: np.ndarray,
+    slope_values: np.ndarray,
+    variance_values: np.ndarray,
+    steepest_rise: int,
+    weights: Sequence[float] = (0.5, 0.35, 0.1, 0.05),
+    verbose: int = 0,
+) -> float:
+    rank_median = np.array(median_values).argsort().argsort()
+    rank_slope = np.abs(slope_values).argsort().argsort()
+    rank_var = np.array(variance_values).argsort().argsort()
+
+    relative_positions = (
+        np.array(positions, dtype=float) - (steepest_rise + 3)
+    )  # Add samples because steepest rise position is sometimes too far left due to filtering
+    bkps_after_event = relative_positions > 0
+    relative_positions[bkps_after_event] = np.nan
+    rank_position = np.abs(relative_positions).argsort().argsort()
+    rank_position[bkps_after_event] += 10  # penalize positions after steepest rise
+
+    if verbose:
+        print("median values", median_values, rank_median)
+        print("slopes", slope_values, rank_slope)
+        print("variances", variance_values, rank_var)
+        print("position", positions, rank_position)
+
+    arr = np.stack([rank_position, rank_median, rank_slope, rank_var])
+    a_weights = np.asarray(list(weights), dtype=float)
+
+    return a_weights @ arr
+
+
+def get_event_baseline(
+    data: np.ndarray,
+    bsl_duration: int,
+    event_num: int,
+    add_points,
+    peak_position: int,
+    positions: np.ndarray,
+    debug: bool = False,
+):
+    """
+    Calculate the baseline and baseline variance for an event in the given data.
+
+    Parameters:
+    - data (np.ndarray): The input data (i.e. the event snippet).
+    - bsl_duration (int): The duration (in points) to consider for baseline calculation.
+    - event_num (int): The index of the event.
+    - add_points (int): The number of additional points to consider.
+    - peak_position (int): The position of the peak relative to start of the event snippet.
+    - positions (np.ndarray): The absolute positions of the events in the main trace.
+    - debug (bool): If True, enables debug mode with additional plots.
+
+    Returns:
+    - baseline (float): The calculated baseline.
+    - bsl_var (float): The calculated baseline variance.
+    - bsl_start (int): The starting index for baseline calculation.
+    - bsl_end (int): The ending index for baseline calculation.
+    """
+
+    previous_peak_present = False
+    if (
+        int(positions[event_num]) - int(positions[event_num - 1]) < add_points
+        and event_num != 0
+    ):
+        previous_peak_present = True
+
+    bsl_limit_factor = 1.5
+    search_end = int(add_points * 2)
+    peak_win_start = add_points // 2
+    win = sc.signal.windows.hann(25)
+
+    penalty = 10
+    trace_start = 0
+
+    if previous_peak_present:
+        penalty = 5
+        trace_start = int(positions[event_num]) - int(positions[event_num - 1])
+        trace_start = min(trace_start, peak_win_start)
+    # check if beginning of baseline is above peak
+    elif np.sum(data[0:peak_win_start] > data[peak_position]) > (peak_win_start / 2):
+        penalty = 5
+        trace_start = int(peak_win_start / 2)
+
+    model = rpt.KernelCPD(kernel="rbf", min_size=2).fit(data[trace_start:search_end])
+    result = model.predict(n_bkps=2)
+    result = [val + trace_start for val in result]
+
+    filtered_data = sc.signal.convolve(data, win, mode="same") / sum(win)
+    gradient = np.gradient(filtered_data)
+    ev_position = np.argmax(gradient[50:300]) + 50
+    cutoff = ev_position - bsl_limit_factor * bsl_duration
+
+    if result[0] < cutoff:
+        if result[1] > peak_position or result[1] < cutoff:
+            result2 = model.predict(pen=penalty)
+            result2 = [val + trace_start for val in result2]
+            onset = result2[np.where(np.array(result2) - peak_position < 0)[0][-1]]
+        else:
+            onset = result[1]
+    else:
+        onset = result[0]
+
+    bsl_end = onset - 10
+    if bsl_end > bsl_duration:
+        bsl_snippet = data[bsl_end - bsl_duration : bsl_end]
+        fit = np.polynomial.polynomial.Polynomial.fit(
+            np.arange(bsl_snippet.shape[0]), bsl_snippet, 1
+        )
+        coef = fit.convert().coef
+
+        if len(coef) > 1 and coef[1] < -0.12 and bsl_duration > 20:
+            bsl_duration = bsl_duration // 2
+            bsl_snippet = data[bsl_end - bsl_duration : bsl_end]
+            fit = np.polynomial.polynomial.Polynomial.fit(
+                np.arange(bsl_snippet.shape[0]), bsl_snippet, 1
+            )
+            coef = fit.convert().coef
+
+            if len(coef) > 1 and coef[1] < -0.12 and bsl_duration > 20:
+                bsl_duration = bsl_duration // 2
+                bsl_snippet = data[bsl_end - bsl_duration : bsl_end]
+                fit = np.polynomial.polynomial.Polynomial.fit(
+                    np.arange(bsl_snippet.shape[0]), bsl_snippet, 1
+                )
+                coef = fit.convert().coef
+
+    return BaselineResult(
+        value=np.median(data[bsl_end - bsl_duration : bsl_end]),
+        var=np.var(data[bsl_end - bsl_duration : bsl_end]),
+        start=bsl_end - bsl_duration,
+        end=bsl_end,
+        duration=bsl_duration,
     )
